Rename package to better reflect the extended domain
